Extract hero image into its own component
